firstcommit

+++ src/main/java/com/helpfinder/repository/CoreSiteReviewsRepository.java
@@ -19,9 +19,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
-import java.security.InvalidParameterException;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.time.Instant;
-import java.time.LocalDate;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
@@ -33,9 +33,9 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Repository;
-import org.apache.commons.io.comparator.LastModifiedFileComparator;  
+import org.apache.commons.io.comparator.LastModifiedFileComparator;
 
+import com.helpfinder.exception.InvalidSiteReviewException;
 import com.helpfinder.model.SiteReview;
 
 @Component
@@ -49,6 +49,7 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 	// this formatter is used to keep the format in which date is stored in file
 	//private SimpleDateFormat dateFormatter = new SimpleDateFormat("dd/MM/yy HH:mm:ss");
 	private DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss.SSSS");
+	private SimpleDateFormat simpleFormatter = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss.SSSS");
 	
 	public CoreSiteReviewsRepository(@Value("${site.review.dir}")String reviewDirectory,
 						@Value("${site.review.dir}")String reviewArchDirectory) {
@@ -61,16 +62,14 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 	 * @param review the review object which should be saved
 	 */
 	@Override
-	public void saveSiteReview(SiteReview review) {
-		Formatter outfile = null;
+	public void saveSiteReview(SiteReview review) throws InvalidSiteReviewException{		
+		//create a new file with user_guid.txt format
+		String fileName = review.getUserId() + "_"+ java.util.UUID.randomUUID().toString() + ".txt";
+		//fileId is the reference to fetch this specific review file	  
+		String fileId = createIfnotExist(reviewDirectory, fileName);
+		 // open file
+	    try(Formatter outfile = new Formatter(fileId)) {		 	 
 
-	      try
-	      {			
-			 //create a new file with user_guid.txt format
-			 String fileName = review.getUserId() + "_"+ java.util.UUID.randomUUID().toString() + ".txt";
-			 //fileId is the reference to fetch this specific review file	  
-			 String fileId = createIfnotExist(reviewDirectory, fileName);	 
-			 outfile = new Formatter(fileId); // open file
 			 /* add the review content format as
 			   * UserId
 			   * 23432 << some user id
@@ -82,41 +81,39 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 			   * review submitted date here
 			   * ReviewId 
 			   * the file path including file name
-			  */
-			  
+			  */			  
 			  outfile.format("UserId\n%d\nTitle\n%s\nComment\n%s\nReviewSubmittedOn\n%s\nReviewID\n%s", review.getUserId(), review.getTitle(), 
 			  review.getComment(), dateFormatter.format(Instant.ofEpochMilli(review.getSubmittedDate().getTime())
-      .atZone(ZoneId.systemDefault())
-      .toLocalDateTime()), fileId);
-			  
-	      }
-	      catch (FileNotFoundException ex)
-	      {
-	          System.err.println("Error opening the file " + ex.getMessage());
-	      }
-	      finally{
-	      	if(outfile != null)
-	      		outfile.flush();
-	      		outfile.close();
-	      }	  
+		      .atZone(ZoneId.systemDefault()).toLocalDateTime()), fileId);			  
+	     }
+	     catch (FileNotFoundException ex) {
+	         System.err.println("Error opening the file " + ex.getMessage());
+	         throw new InvalidSiteReviewException(ex.getMessage());	
+	     }	       
 	}	
 	/**
 	*creates a file if does not exist for the given file path
 	*@param filePath the file path 
 	*@param fileName the file name
 	*@return String the full file path 
+	 * @throws InvalidSiteReviewException 
 	*/
-	private String createIfnotExist(String filePath, String fileName) {
+	private String createIfnotExist(String filePath, String fileName) throws InvalidSiteReviewException {	
+		//construct the directory path where the file needs to be created
 		File file = new File(FileSystems.getDefault().getPath("").toAbsolutePath() + File.separator + filePath);
+		//if directory does not exist then create it
 		if(!file.exists()){
 			file.mkdirs();
 		}
+		// create the file instance for the file path
 		file = new File(file.getPath() + File.separator +fileName);
+		// if file does not exist then create the file
 	    if(!file.exists()){	      
 	        try {
 	         	file.createNewFile();
 			} catch (IOException e) {
-				System.err.println("Error creating file " + e.getMessage());			
+				System.err.println("Error creating file " + e.getMessage());		
+				throw new InvalidSiteReviewException(e.getMessage());	
 			}
 	    }
 	    return file.getPath();
@@ -125,18 +122,12 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 	/**
 	 * gets a list of reviews up to the limit sorted by date descending of the file created
 	 * @param limit the number of reviews to return
-	 * @param sortDateDesc boolean value indicating if the reviews fetched are to be sorted descending based on reviewed date   
-	 * @throws  
+	 * @param sortDateDesc boolean value indicating if the reviews fetched are to be sorted descending based on reviewed date  
+	 * @throws InvalidSiteReviewException 
+	   
 	 */
 	@Override
-	public List<SiteReview> getSiteReview(int limit, boolean sortDateDesc) {
-		// can only get max of 20 files
-		if(limit < 1)
-			throw new InvalidParameterException("limit should be greater than 0");
-		// process only to max limit
-		int maxLimit = 20;
-		if(limit > maxLimit)
-			limit = maxLimit;
+	public List<SiteReview> getSiteReview(int limit, boolean sortDateDesc) throws InvalidSiteReviewException {
 		
 		// list of reviews to return
 		List<SiteReview> reviews = new ArrayList<SiteReview>();
@@ -145,6 +136,8 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 			File directory = new File(FileSystems.getDefault().getPath("").toAbsolutePath()+ File.separator+ reviewDirectory);
 			// get the file list
 			File[] files = directory.listFiles();
+			if(files == null || files.length == 0)
+				return reviews;
 			// sort based on last modified date
 	        Arrays.sort(files, sortDateDesc ?  LastModifiedFileComparator.LASTMODIFIED_COMPARATOR : LastModifiedFileComparator.LASTMODIFIED_REVERSE);
 	        //SiteRevie list to return
@@ -174,30 +167,29 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 						case 5:
 						 		review.setComment(curLine);
 						 		break;
-						case 7:  review.setSubmittedDate(java.util.Date.from(LocalDate.parse(curLine, dateFormatter).atStartOfDay()
-      								.atZone(ZoneId.systemDefault())
-      								.toInstant()));
-					
+						case 7: review.setSubmittedDate(simpleFormatter.parse(curLine));
 					 			break;
 					 	case 9:
 						 		review.setReviewId(curLine);
 					 			break;
 					 	default: break;
 					}
-					//add to list
-					reviews.add(review);
+					
 				   //get the next line
 					curLine = reader.readLine();
 					//increment the line
-					line++;
+					line++;					
 				 }
-				 reader.close();	   
-				 // decrease the limit
-				 limit--;   	        	  
+				 //add to list
+				reviews.add(review);
+				 reader.close();	  
+				  // decrease the limit
+				 limit--;   	  
 			}
 		}
-		catch(IOException ex) {
+		catch(ParseException | IOException ex) {
 			 System.err.println("Error fetching files " + ex.getMessage());
+			 throw new InvalidSiteReviewException(ex.getMessage());
 		}
 		return reviews;
 	}
@@ -205,17 +197,23 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 	/**
 	 * Moves the review to archived folder
 	 * @param id the source file name with the full path which should be archived
+	 * @throws InvalidSiteReviewException 
 	 * @throws IOException 
 	 */
 	@Override
-	public void archiveReview(String reviewId) {
+	public void archiveReview(String reviewId) throws InvalidSiteReviewException {
 	
 		Formatter outfile = null;
 		try {
 			// create source and destination path
-			Path source = Paths.get(reviewId);			
+			Path source = Paths.get(reviewId);
+			File sourceFile = new File(reviewId);
+			if(!sourceFile.exists() || !sourceFile.isFile())
+				throw new InvalidSiteReviewException("Review Id is Invalid");
+										
 			String destinationFilePath = createIfnotExist(reviewArchDirectory , source.getFileName().toString());
 			 Path destination = Paths.get(destinationFilePath);
+			 
 			// move the file
 			Files.move(source, destination , StandardCopyOption.REPLACE_EXISTING);
 			// append in the file the reviewed date
@@ -227,6 +225,7 @@ public class CoreSiteReviewsRepository implements SiteReviewsRepository{
 		}
 		catch(IOException ex) {
 			System.err.println("Error archiving the files " + ex.getMessage());
+			throw new InvalidSiteReviewException(ex.getMessage());
 		}
 		finally	{
 			if(outfile != null){
